test(examples): extract helpers so the plugin examples can be linted
Four packages under examples/ each declared the same symbol behind both
consumer and provider build tags: dir in examples/plugin and
examples/avro, generateMattMessage and parseMattMessage in
examples/plugin, and User in the examples root package. The tags meant
each pair of files never compiled together in a normal test run, so
the duplication was invisible. A lint pass covering both tags compiles
them together, and it failed with redeclaration errors before reaching
a single linter.

That left roughly 3,600 lines of examples unlinted, which is why the
build-tags configuration could not be turned on. Leaving any one of
the four packages unfixed would have made the build-tags config break
mise run lint repo-wide, blocking every later lint-ratchet task rather
than just the examples.

Where the two copies of a declaration had diverged, the consumer's
copy wins:

- generateMattMessage/parseMattMessage: the consumer's version embeds
  the trailing "\n" that the provider wrote as a second, separate
  conn.Write call; that second write is now redundant and is removed,
  leaving the exact same bytes on the wire.
- User: the consumer's struct tags are load-bearing (read by
  reflection via BodyMatch/AsType to build the pact body), while the
  provider constructs the struct with explicit field values and never
  reads its own tags - so the provider's now-removed tag was already
  inert, and taking the consumer's corrects an inconsistency it had
  with its own field literal.

Where the two copies were already identical (dir, in both plugin and
avro), the declaration moves as-is.

Each surviving declaration moves to an untagged helpers_test.go in its
package, and the lint configuration now covers both tags. No new
linters are enabled here; this only makes the examples visible to the
ones that follow.
