+++ .golangci.yml
@@ -1,6 +1,24 @@
 version: "2"
+run:
+  build-tags:
+    - consumer
+    - provider
 linters:
   exclusions:
     paths:
       # lib.go only contains cgo annotations
       - internal/native/lib.go
+    rules:
+      # TEMPORARY. Adding consumer/provider to run.build-tags brings ~3,600
+      # previously uncompiled lines under examples/ into scope, and the linters
+      # already on by default immediately report 21 findings there: 18 errcheck,
+      # one govet hostport and two staticcheck SA1019 deprecations. Fixing them
+      # here would pull the error-handling and safety work forward out of the
+      # commits that own it, so the examples are excused from those three
+      # linters until it lands. Remove this rule once the safety pass fixes the
+      # last of them.
+      - path: ^examples/
+        linters:
+          - errcheck
+          - govet
+          - staticcheck
