+++ examples/avro/helpers_test.go
@@ -1,0 +1,5 @@
+package avro
+
+import "os"
+
+var dir, _ = os.Getwd()
