+++ examples/consumer_v2_test.go
@@ -157,13 +157,6 @@ func TestConsumerV2AllInOne(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-type User struct {
-	ID       int    `json:"id" pact:"example=27"`
-	Name     string `json:"name" pact:"example=Billy"`
-	LastName string `json:"lastName" pact:"example=Sampson"`
-	Date     string `json:"datetime" pact:"example=2020-01-01'T'08:00:45,format=yyyy-MM-dd'T'HH:mm:ss,generator=datetime"`
-}
-
 // Pass in test case
 
 var test = func() func(config consumer.MockServerConfig) error {
