+++ examples/helpers_test.go
@@ -1,0 +1,8 @@
+package main
+
+type User struct {
+	ID       int    `json:"id" pact:"example=27"`
+	Name     string `json:"name" pact:"example=Billy"`
+	LastName string `json:"lastName" pact:"example=Sampson"`
+	Date     string `json:"datetime" pact:"example=2020-01-01'T'08:00:45,format=yyyy-MM-dd'T'HH:mm:ss,generator=datetime"`
+}
