+++ examples/provider_test.go
@@ -239,10 +239,3 @@ func startServer() {
 
 	l.Fatal(http.ListenAndServe("127.0.0.1:8111", mux))
 }
-
-type User struct {
-	ID       int    `json:"id" pact:"example=27"`
-	Name     string `json:"name" pact:"example=billy"`
-	LastName string `json:"lastName" pact:"example=Sampson"`
-	Date     string `json:"datetime" pact:"example=2020-01-01'T'08:00:45,format=yyyy-MM-dd'T'HH:mm:ss,generator=datetime"`
-}
